.github/scripts/requirements.txt
requests
python-dotenv
linkedin-api